remove Comparable into Text, ComparableText class.
