+++ src/main/java/org/cactoos/text/Base64Text.java
@@ -70,8 +70,4 @@ public final class Base64Text implements Text {
         ).asString();
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
 }
