+++ src/main/java/org/cactoos/text/FormattedText.java
@@ -171,9 +171,4 @@ public final class FormattedText implements Text {
         return out.toString();
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
-
 }
