+++ src/main/java/org/cactoos/text/JoinedText.java
@@ -101,9 +101,4 @@ public final class JoinedText implements Text {
         return joint.toString();
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
-
 }
