+++ src/main/java/org/cactoos/text/NormalizedText.java
@@ -62,10 +62,5 @@ public final class NormalizedText implements Text {
         return new TrimmedText(this.origin).asString().replaceAll("\\s+", " ");
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
-
 }
