+++ src/main/java/org/cactoos/text/RepeatedText.java
@@ -75,9 +75,4 @@ public final class RepeatedText implements Text {
         return out.toString();
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
-
 }
