+++ src/main/java/org/cactoos/text/ReplacedText.java
@@ -68,10 +68,5 @@ public final class ReplacedText implements Text {
         return this.origin.asString().replace(this.needle, this.replacement);
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
-
 }
