+++ src/main/java/org/cactoos/text/ReversedText.java
@@ -54,9 +54,4 @@ public final class ReversedText implements Text {
         return new StringBuilder(this.origin.asString()).reverse().toString();
     }
 
-    @Override
-    public int compareTo(final Text text) {
-        return new UncheckedText(this).compareTo(text);
-    }
-
 }
