+++ src/test/java/org/cactoos/text/TextOfTest.java
@@ -171,20 +171,6 @@ public final class TextOfTest {
     }
 
     @Test
-    public void comparesWithASubtext() throws Exception {
-        MatcherAssert.assertThat(
-            "Can't compare sub texts",
-            new TextOf(
-                "here to there"
-            ).compareTo(
-                // @checkstyle MagicNumberCheck (1 line)
-                new SubText("from here to there", 5)
-            ),
-            Matchers.is(0)
-        );
-    }
-
-    @Test
     public void readsStringBuilder() throws IOException {
         final String starts = "Name it, ";
         final String ends = "then it exists!";
